// src/compiler/meta/types.d.ts
export interface Script {
    /**
     * TypeScript script file
     */
    src: string;
    /**
     * The type of source code
     */
    type: 'server' | 'client' | 'shared';
    /**
     * When the script file type is "client", this setting controls whether the file is saved on the clients' hard drive
     */
    cache?: boolean;
    /**
     * If set to "false", compatibility checks are skipped
     */
    validate?: boolean;
}